requirements.txt
# Core dependencies
tiktoken>=0.5.0
requests>=2.31.0
google-genai>=0.2.0

# Testing
hypothesis>=6.92.0
pytest>=7.4.0

# Utilities
python-dateutil>=2.8.2
